Decode The Day
Can you guess today is which day of the year? Let’s check your capability of finding out the correct day of the year when you are given the DATE!. You are given a Date as an input and your task is to determine the day of the year which falls on the given Date. So, get ready to go on a date with the DATE! (Assume that the year is non leap year).

Input Format

The first line will give T, the number of testcases.
Each testcase will give the date to decode in DDMMM format.
Constraints

1<=T<=500
Output Format

For Each testcase, output the count of the day of the given date from the start of the year.

Sample Input 0

2
10JAN
02FEB
Sample Output 0

10
33