Pass The Soan Papdi
It's Diwali time, that means fun time! Enjoying with the family, playing interesting games and eating a lot of delicious sweets. And the flavor of the season is always our very own favorite soan papadi! That we all love to eat.

Imagine you and your family members are playing an amazing game of passing soan papadi. In this game your n members of the family are standing in a line labeled from 1 to n. The first person in the line is holding the packet of soan papdi initially. Every second, the person holding the packet passes it to the next person standing in the line. Once the packet reaches the end of the line, the direction changes, and people continue passing the packet in the opposite direction.

For example, once the packet of soan papdi reaches the nth person they pass it to the n - 1th person, then to the n - 2th person and so on.

Given the two positive integers n and time, return the position of the person who wins the game and get the packet of soan papdi after time seconds.

Input Format

The first line will give T, number of testcases.
Every testcase will provide two positive integers n and time.
Constraints

1<=T<=1000
2<=n<=1000
1<=time<=1000
Output Format

The position of the person who wins the game

Sample Input 0

1
4 5
Sample Output 0

2
Explanation 0

People pass the soan papdi in the following way: 1 -> 2 -> 3 -> 4 -> 3 -> 2. Afer five seconds, the soan papdi is given to the 2nd person.

Sample Input 1

1
3 2
Sample Output 1

3
Explanation 1

People pass the soan papdi in the following way: 1 -> 2 -> 3.

Afer two seconds, the soan papdi is given to the 3rd person.