Player Recruitment
You have an array called 'indian_team' that contains the jersey numbers of players currently in the squad. The goal is to intelligently pick k distinct players (represented as positive integers) who are not presently in the team. Each player's cost is determined by the number on their jersey. Your objective is to choose k new distinct players, not in the current team, in a way that minimizes the total cost acquired by adding them to the team. Provide the minimum cost required to add k new players.

Input Format

First line will be T, the number of testcases.
The first line of each testcase will be n, the number of players already in team and k (number of players to add to team).
The second line of each testcase will be array indian_team of n elements.
Constraints

1<=T<=10^3
1<=K<=10^8
1<=n<=10^5
1<=indian_team[i]<=10^9
Output Format

For each testcase output the cost added to total cost by adding k new players.

Sample Input 0

1
5 2
1 4 25 10 25
Sample Output 0

5
Explanation 0

The two unique positive integers that do not appear in array which we append are 2 and 3. The resulting sum of nums is 1 + 4 + 25 + 10 + 25 + 2 + 3 = 70, which is the minimum. The sum of the two integers appended is 2 + 3 = 5, so we return 5.

Sample Input 1

1
2 6
5 6
Sample Output 1

25
Explanation 1

The six unique positive integers that do not appear in array which we append are 1, 2, 3, 4, 7, and 8. The resulting sum of nums is 5 + 6 + 1 + 2 + 3 + 4 + 7 + 8 = 36, which is the minimum. The sum of the six integers appended is 1 + 2 + 3 + 4 + 7 + 8 = 25, so we output 25.