The unimatrix challange
Its game time! We have an interesting game for you and your friend, imagine that you have a mxn number matrix, you give an integer x to your friend. He or she can add or subtract the number x from any element in the matrix. The task for your friend is to create a uni value matrix which means all the elements of the matrix should be equal or should have equal value by performing the operations of addition and subtraction on elements of the matrix using the number x. You need to output the minimum number of operations performed by your friend to make the matrix uni-value. If it is not possible to make one, output -1.

Input Format

First line will be T, the number of testcases.

The first line of each testcase will give m and n.

The next lines will provide the matrix.

Constraints

1<=T<=10^3

1<=m,n<=25

1<=x<=10

-100000<=matrix[i][j]<=100000

Output Format

You need to output the minimum number of operations performed by your friend to make the matrix uni-value. If it is not possible to make one, output -1.

Sample Input 0

1
2 2 2
2 4
6 8
Sample Output 0

4
Sample Input 1

1
2 2 1
1 5
2 3
Sample Output 1

5