K-Diff Hacks
You are a hacker who has infiltrated a secret facility that stores an array of integers nums and an integer k. You want to find out how many unique k-diff pairs are hidden in the array.

A k-diff pair is a pair of numbers (nums[i], nums[j]) that satisfy the following conditions:

0 <= i, j < nums.length i != j.
The absolute difference between nums[i] and nums[j] is equal to k.
The absolute difference is calculated as |val|, which means the positive value of val.
How many unique k-diff pairs can you discover in the array?

Input Format

The first line will give T, the number of testcases.
The first line of each testcase will give n, size of nums and k.
The second line of each testcase will give array nums
Constraints

1<=T<=1000
1 <= nums.length <= 10^4
-10^7 <= nums[i] <= 10^7
0 <= k <= 10^7
Output Format

For each testcase, Print the number of unique k-diff pairs can be discovered in the array?

Sample Input 0

1
5 2
3 1 4 1 5
Sample Output 0

2
Explanation 0

There are two 2-diff pairs in the array, (1, 3) and (3, 5). Although we have two 1s in the input, we should only print the number of unique pairs.

Sample Input 1

1
5 1
1 2 3 4 5
Sample Output 1

4
Explanation 1

There are four 1-diff pairs in the array, (1, 2), (2, 3), (3, 4) and (4, 5).