Last stone weight
You are given an array of integers stones where stones[i] is the weight of the ith stone.

We are playing a game with the stones. On each turn, we choose the heaviest two stones and smash them together. Suppose the heaviest two stones have weights x and y with x <= y. The result of this smash is:

If x == y, both stones are destroyed, and If x != y, the stone of weight x is destroyed, and the stone of weight y has new weight y - x. At the end of the game, there is at most one stone left.

Print the weight of the last remaining stone. If there are no stones left, Print 0.

Input Format

The first line will give T, the number of testcases.
The first line of each testcase will give n the number of stones.
The second line of each testcase will give array stones where stones[i] is the weight of the ith stone..
Constraints

1 <= T <= 1000
1 <= stones.length <= 30
1 <= stones[i] <= 1000
Output Format

For each testcase print the weight of the last remaining stone. If there are no stones left, Print 0.

Sample Input 0

1
6
2 7 4 1 8 1
Sample Output 0

1
Explanation 0

We combine 7 and 8 to get 1 so the array converts to [2,4,1,1,1] then,
we combine 2 and 4 to get 2 so the array converts to [2,1,1,1] then,
we combine 2 and 1 to get 1 so the array converts to [1,1,1] then,
we combine 1 and 1 to get 0 so the array converts to [1] then that's the value of the last stone.