Find the score
You are given an array nums consisting of positive integers.

Starting with score = 0, apply the following algorithm:

Choose the smallest integer of the array that is not marked. If there is a tie, choose the one with the smallest index.
Add the value of the chosen integer to score.
Mark the chosen element and its two adjacent elements if they exist.
Repeat until all the array elements are marked.
Print the score you get after applying the above algorithm.

Input Format

The first line will give T, the number of testcases.
The first line of each testcase will give the size of array nums.
The second line of each testcase will give array nums.
Constraints

1 <= T <= 1000
1 <= nums.length <= 10^5
1 <= nums[i] <= 10^6
Output Format

For each testcase print the score you get after applying the above algorithm.

Sample Input 0

1
6
2 1 3 4 5 2
Sample Output 0

7
Explanation 0

We mark the elements as follows:

1 is the smallest unmarked element, so we mark it and its two adjacent elements: [2,1,3,4,5,2].
2 is the smallest unmarked element, so we mark it and its left adjacent element: [2,1,3,4,5,2].
4 is the only remaining unmarked element, so we mark it: [2,1,3,4,5,2].
Our score is 1 + 2 + 4 = 7.

Sample Input 1

1
6
2 3 5 1 3 2
Sample Output 1

5
Explanation 1

We mark the elements as follows:

1 is the smallest unmarked element, so we mark it and its two adjacent elements: [2,3,5,1,3,2].
2 is the smallest unmarked element, since there are two of them, we choose the left-most one, so we mark the one at index 0 and its right adjacent element: [2,3,5,1,3,2].
2 is the only remaining unmarked element, so we mark it: [2,3,5,1,3,2].
Our score is 1 + 2 + 2 = 5.