Time needed to buy tickets
There are n people in a line queuing to buy tickets, where the 0th person is at the front of the line and the (n - 1)th person is at the back of the line.

You are given a 0-indexed integer array tickets of length n where the number of tickets that the ith person would like to buy is tickets[i].

Each person takes exactly 1 second to buy a ticket. A person can only buy 1 ticket at a time and has to go back to the end of the line (which happens instantaneously) in order to buy more tickets. If a person does not have any tickets left to buy, the person will leave the line.

Print the time taken for the person at position k (0-indexed) to finish buying tickets.

Input Format

The first line will give T, the number of testcases.
The first line of each testcase will give n and k.
The second line will give 0-indexed integer array tickets of length n where the number of tickets that the ith person would like to buy is tickets[i].
Constraints

1 <= T <= 1000
n == tickets.length
1 <= n <= 100
1 <= tickets[i] <= 100
0 <= k < n
Output Format

For each testcase, print the time taken for the person at position k (0-indexed) to finish buying tickets.

Sample Input 0

1
3 2
2 3 2
Sample Output 0

6
Explanation 0

In the first pass, everyone in the line buys a ticket and the line becomes [1, 2, 1].
In the second pass, everyone in the line buys a ticket and the line becomes [0, 1, 0].
The person at position 2 has successfully bought 2 tickets and it took 3 + 3 = 6 seconds.

Sample Input 1

1
4 0
5 1 1 1
Sample Output 1

8
Explanation 1

In the first pass, everyone in the line buys a ticket and the line becomes [4, 0, 0, 0].
In the next 4 passes, only the person in position 0 is buying tickets.
The person at position 0 has successfully bought 5 tickets and it took 4 + 1 + 1 + 1 + 1 = 8 seconds.