Game of Circle
There are n friends that are playing a game. The friends are sitting in a circle and are numbered from 1 to n in clockwise order. More formally, moving clockwise from the ith friend brings you to the (i+1)th friend for 1 <= i < n, and moving clockwise from the nth friend brings you to the 1st friend.

The rules of the game are as follows:

Start at the 1st friend.
Count the next k friends in the clockwise direction including the friend you started at. The counting wraps around the circle and may count some friends more than once.
The last friend you counted leaves the circle and loses the game.
If there is still more than one friend in the circle, go back to step 2 starting from the friend immediately clockwise of the friend who just lost and repeat.
Else, the last friend in the circle wins the game.
Given the number of friends, n, and an integer k, print the winner of the game.

Input Format

The first line will give T, the number of testcases.
Each testcase will give 2 numbers, n and k.
Constraints

1 <= T <= 1000
1 <= k <= n <= 500
Output Format

For each testcase, Print the winner of the game.

Sample Input 0

1
5 2
Sample Output 0

3
Explanation 0

Here are the steps of the game:

Start at friend 1.
Count 2 friends clockwise, which are friends 1 and 2.
Friend 2 leaves the circle. Next start is friend 3.
Count 2 friends clockwise, which are friends 3 and 4.
Friend 4 leaves the circle. Next start is friend 5.
Count 2 friends clockwise, which are friends 5 and 1.
Friend 1 leaves the circle. Next start is friend 3.
Count 2 friends clockwise, which are friends 3 and 5.
Friend 5 leaves the circle. Only friend 3 is left, so they are the winner.
Sample Input 1

1
6 5
Sample Output 1

1
Explanation 1

The friends leave in this order: 5, 4, 6, 2, 3. The winner is friend 1.