Longest Zig Zag path
You are given an array nums of n elements. Construct a binary search tree from the given array (just like previous days questions).then print the length of longest zig-zag path

A ZigZag path for a binary tree is defined as follow:

Choose any node in the binary tree and a direction (right or left).
If the current direction is right, move to the right child of the current node; otherwise, move to the left child.
Change the direction from right to left or from left to right.
Repeat the second and third steps until you can't move in the tree.
Zigzag length is defined as the number of nodes visited - 1. (A single node has a length of 0).
Print the longest ZigZag path contained in that tree.

Input Format

The first line will give T, the number of testcases.
The first line of each testcase will give n, the size of array nums.
The second line of each testcase will give the array nums.
Constraints

1 <= T <= 1000
1 <= n <= 100
-10^3 <= nums[i] <= 10^3
Output Format

For each testcase, print the length of longest zigzag path contained in the tree constructed by given array nums.

Sample Input 0

1
10
1 10 2 9 3 8 4 7 5 6
Sample Output 0

9